pyproject.toml
[project]
name = 'cvxball'
version = "0.0.0"
description = "Fast computation of the smallest enclosing sphere"
authors = [{name='Thomas Schmelzer', email= 'thomas.schmelzer@gmail.com'}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cvxpy-base>=1.6.0",
    "numpy>=2",
    "clarabel>=0.10.0",
    "numpy-flight>=0.0.14",
]

[project.urls]
repository = "https://github.com/cvxgrp/cvxball"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pre-commit>=4.0.1",
    "kaleido==1.0.0",
    "mosek==11.0.25",
    "matplotlib>=3.10.0",
    "plotly>=5.24.1",
    "marimo>=0.11.6",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "*__init__.py"
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cvx"]

[tool.deptry.per_rule_ignores]
DEP002 = ["clarabel","numpy-flight"]
DEP003 = ["cvx"]

[tool.bandit]
exclude_dirs = ["tests"]
exclude = ["B104"]

[project.scripts]
#server = "cvx.ball.server:BallServer.start"
#client = "example.client:main"
demo = "example.client:main"

[tool.deptry.package_module_name_map]
cvxpy-base = ["cvxpy"]
numpy-flight = ["flight"]
